add additional data sources + refactor and test snooty sources
